+++ CHANGELOG.md
@@ -5,6 +5,17 @@
 ### 🐛 Bug Fixes
 
 - Check for duplicate operation IDs in workflows (#59) by @mbrancato
+- Add Closer interface for Modules to cleanup resources after workflow run (#65) by @mbrancato
+
+### 💼 Other
+
+- Sanitize logs with Google API Errors when wrapped (#61) by @mbrancato
+
+### 📚 Documentation
+
+- Add Cloud SQL grant limitation docs for `cloudsqlsuperuser` (#62) by @mbrancato
+- Normalize Cloud SQL naming to match official documentation (#63) by @mbrancato
+- Normalize PostgreSQL naming to match official documentation (#64) by @mbrancato
 
 ## [0.1.9] - 2026-03-27